% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comp_table.R
\name{comp_table}
\alias{comp_table}
\title{Creates a detailed table for compensation
Containing the hours worked on a specific project within a specified timeframe
Uses lazy evaluation}
\usage{
comp_table(
  data,
  agg_by = c("Month", "Task", "Project"),
  min_date = NULL,
  max_date = NULL,
  proj_name = NULL,
  client_name = NULL,
  available_comp = NULL,
  verbose = FALSE,
  date = Date,
  hours = Hours,
  compensation = Compensation,
  project = Project,
  task = Task,
  client = Client
)
}
\arguments{
\item{data}{A data.frame containing project data}

\item{agg_by}{NULL if detailed item-by-item output is desired, a character vector with one or more of "Month", "Task", "Project" otherwise}

\item{proj_name}{If output for only one project is desired, the name of the project}

\item{client_name}{If output for only one client is desired, the name of the client}

\item{available_comp}{If working on a retainer basis, the amount of money available at the start of the project}

\item{date}{Name of the Date column in your data set}

\item{hours}{Name of the Hours column in your data set}

\item{compensation}{Name of the Compensation column in your data set}

\item{project}{Name of the Project column in your data set}

\item{task}{Name of the column containing a Short description of the task carried out in specified time frame}

\item{client}{Name of the Client column in your data set}

\item{mindate}{A date in the format as.Date("yyyy-mm-dd")}

\item{maxdate}{A date in the format as.Date("yyyy-mm-dd")}

\item{table}{should a kable be returned or the raw data.frame?}
}
\value{
A formatted table of compensation grouped by either combination of month, project and client.
}
\description{
Creates a detailed table for compensation
Containing the hours worked on a specific project within a specified timeframe
Uses lazy evaluation
}
\examples{
comp_table(data = timesheet, client_name = "Client B")
comp_table(data = timesheet, agg_by = "Task", min_date = as.Date("2022-01-01"), max_date = as.Date("2022-04-30"))
comp_table(data = timesheet, agg_by = "Task", min_date = as.Date("2022-01-01"), max_date = as.Date("2022-04-30"), proj_name = "Project 1")
}
