% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_invoice.R
\name{render_invoice}
\alias{render_invoice}
\title{Renders and invoice
By default using the template contained in the package}
\usage{
render_invoice(
  data,
  client,
  address,
  proj_name,
  inv_number,
  with,
  discount,
  VAT = "20\%",
  currency = "Euro",
  template = system.file("inst/rmarkdown/templates/invoice/skeleton", "skeleton.Rmd",
    package = "admintools"),
  filename = paste0(Sys.Date(), "-SENDER-RECIPIENT-NUMBER.pdf"),
  dir = "C:/Users/jonas/Documents/testinvoices",
  verbose = FALSE
)
}
\arguments{
\item{data}{A data.frame containing project data as returned by report_single}

\item{proj_name}{Name of the project you want to write the invoice for. One of either proj_name or client_name is required.}

\item{template}{Two templates are provided: "organization.Rmd" for organizational clients and "individual.Rmd" for individuals. If these don't meet your needs you can point to your own template.}

\item{clientlist}{A data.frame containing names & contact information for your clients}

\item{client_name}{Name of the client you want to write the invoice for.  One of either proj_name or client_name is required.}
}
\value{
A formatted invoice.
}
\description{
Renders and invoice
By default using the template contained in the package
}
\examples{
render_invoice(
  data = timesheet,
  client = filter(addresses, Client == "Client B"),
  address = data.frame(filter(addresses, Client == "talynsight")),
  proj_name = "Project 1",
  inv_number = "10098",
  with = "PI Buck Mulligan, PhD",
  discount = "10\%"
  )
}
