% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_overall.R
\name{report_overall}
\alias{report_overall}
\title{Creates an overview table for compensation
Can be grouped by either combination of month, project and client
Follows lazy evaluation}
\usage{
report_overall(
  data,
  availablecomp = 11683,
  minmonth = 1,
  maxmonth = 4,
  cpt = "Overview over total monthly compensation and remaining budget",
  Month = Month,
  Hours = Hours,
  Compensation = Compensation,
  Project = NULL,
  Client = NULL
)
}
\arguments{
\item{data}{A data.frame/tibble/tidytable/matrix containing project data}

\item{availablecomp}{Either the available funding at the beginning of the project or NULL if invoice is to be given after the project.}

\item{minmonth}{A number, the first month to consider}

\item{maxmonth}{A number, the last month to consider}

\item{Month}{Name of the Month column in your data set}

\item{Hours}{Name of the Hours column in your data set}

\item{Compensation}{Name of the Compensation column in your data set}

\item{Project}{Name of the Project column in your data set}

\item{Client}{Name of the Client column in your data set}
}
\value{
A formatted table of compensation grouped by either combination of month, project and client.
}
\description{
Creates an overview table for compensation
Can be grouped by either combination of month, project and client
Follows lazy evaluation
}
\examples{
report_overall(df)
report_overall(df, Project = Project)
report_overall(df, Project = Project, minmonth = 2, maxmonth = 3)
}
