% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_single.R
\name{report_single}
\alias{report_single}
\title{Creates a detailed table for compensation
Containing the hours worked on a specific project within a specified timeframe
Follows lazy evaluation}
\usage{
report_single(
  data,
  mindate,
  maxdate,
  proj_name,
  pi = NULL,
  cpt = NULL,
  Date = Date,
  Hours = Hours,
  Compensation = Compensation,
  Project = Project,
  Task = Task
)
}
\arguments{
\item{data}{A data.frame/tibble/tidytable/matrix containing project data}

\item{mindate}{A date in the format as.Date("yyyy-mm-dd")}

\item{maxdate}{A date in the format as.Date("yyyy-mm-dd")}

\item{Date}{Name of the Date column in your data set}

\item{Hours}{Name of the Hours column in your data set}

\item{Compensation}{Name of the Compensation column in your data set}

\item{Project}{Name of the Project column in your data set}

\item{Task}{Name of the column containing a Short description of the task carried out in specified time frame}

\item{Client}{Name of the Client column in your data set}
}
\value{
A formatted table of compensation grouped by either combination of month, project and client.
}
\description{
Creates a detailed table for compensation
Containing the hours worked on a specific project within a specified timeframe
Follows lazy evaluation
}
\examples{
report_single(df)
report_single(df, mindate = as.Date("2022-01-01"), maxdate = as.Date("2022-04-30"), proj_name = "FIMADIA", pi = "Naomi Lange")
report_single(df, Project = Project, minmonth = 2, maxmonth = 3)
}
